lad: Compile lad_dra7xx as app for Android
lad_dra7xx is being launched as an Android service. The
service status is tracked based on the pid of the binary
run when the service is launched.

In daemon mode, the pid that Android keeps track of is that
of the parent process that started the daemon.  So, the
lad_dra7xx service status is not properly shown.

The lad_dra7xx will be compiled in app mode for Android to
avoid the problem described above.  The lad_dra7xx will
continue as a daemon for Linux.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>