lad: Skip daemon running check in Android
The Android service start already takes care of checking
if the service is already running, so it's not required
to check again if the daemon is running or not.

It's recommended to add an Android service in the init.rc
script file for the lad_dra7xx.  For instance:

  service lad_dra7xx /system/bin/lad_dra7xx -g
      class main
      user media
      group system
      oneshot

The service can be set up to automatically launch when
Android boots or manually started later:

  $ start lad_dra7xx

Directly running the lad_dra7xx executable in Android is
discouraged as the capability to enforce that only one
instance is running is lost.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>