Linux: Update user AF_RPMSG define for 4.7+ kernels
IPC's Linux user space support contains a copy of a few
kernel-defined structs and definitions including the socket
address family used for rpmsg-based sockets - AF_RPMSG. This
definition is always 'one more' than the last address family
defined by the mainline kernel's socket.h.

Prior to Linux 3.9, this was aways 40. Couple of new address
families were introduced in 4.6 (AF_KCM) and 4.7 (AF_QIPCRTR),
and as a result, we needed to set AF_RPMSG to 43 post 4.7 kernel.
The user space header has been updated to reflect on the kernel
version and appropriately assign the value of AF_RPMSG.

Do note that a specific value of AF_RPMSG can always be provided
through the products.mak when KERNEL_INSTALL_DIR is not set.

Signed-off-by: Suman Anna <s-anna@ti.com>