Watchdog: Add Config to disable Watchdog
Add a new module config to the Watchdog module that
allows to disable the module. The Watchdog
module is automatically included when Deh module is
included in an image. This parameter allows the user
to disable the watchdog while still using the Deh
module and allows this to be done from a config param
at the application level on the BIOS-side application.

The following can be added to the application config
file to disable Watchdog module (if desired) when including
Deh module.

    WD = xdc.useModule('ti.deh.Watchdog');
    WD.disableWatchdog = true;

There are other ways to disable the watchdog from Linux/Android
that already exist. This method just provides additional
flexibility to the user.

The default value is false if this parameter is not set.
So, if this parameter is not set, the behavior is the same as
before this patch.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>