QNX: NameServer_delete in NameServer daemon leaks memory
Remove an invalid assertion that the name list must be empty.
This is not true and the instance might still be in use by
another client (object is reference counted). Add a while loop
to delete each entry on the name list. Destroy the object's
gate (was not being done). Remove the macro to destruct a
list object (not needed).

[Ported from commits 8a52a6a5fa20ac028158a3ad346da46555ce45b7
    "SDOCM00115434 NameServer_delete in LAD daemon leaks memory"
    and 6c4799151327f2c83b9f253ea25c52bef4e10803
    "Remove NameServer instance from module list in NameServer_delete()"]

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>