Watchdog: Add Watchdog_restore for DRA7XX DSP
Watchdog code is not currently adding the Watchdog_restore
callback for DRA7XX DSP. When power management (suspend/resume)
is enabled for the DSP, the restore hook should be added
so that the values of the timer are proper upon resume.

Without this patch, when power management is enabled for the DSP,
the watchdog timeout is not proper after resuming from suspend.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>