QNX: Build: Fix Build for Newer QNX Versions on Windows
When building on Windows, the overriding of the usemsg
command in quiet.mk was causing the build to fail. It
appears that usemsg isn't always defined when building
with newer QNX versions and the overriding of the command should
be done conditionally.

This patch changes quiet.mk to only update the usemsg
command if it is defined.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>