QNX: NameServer Daemon: Check that Handles passed from User are Valid
Check that the handle being passed from the user is a valid
handle that the daemon has previously allocated. This prevents
the resource manager from crashing by accessing a stale handle
that the user has passed (for instance in some error recovery
scenarios).

This patch adds the checking for function calls which are accessible
through a devctl call.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>