Android: update version.h include path for AF_RPMSG handling
Update include path of version.h for current Android setup.
Previously (before O) we used to update the socket.h bionic header with
AF_RPMSG definitions, so the KERNEL_VERSION check was not exercized.

Now the kernel-headers of dra7xx in Android repo is updated to
reflect the correct Kernel version and can be used to check
the kernel version.

For older Android versions that do not have version.h in
kernel-headers, AF_RPMSG can still be taken from the socket.h
bionic header if it is updated to include AF_RPMSG.

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>