Vayu: NotifySetup: Incorrect Index in NotifySetup_unplugHwi
When checking the reference count before deleting the Hwi,
the code was always checking against index 0, instead of
the mbxIdx. This lead to the Hwi being deleted, even when
there were still references. Then, even if NotifySetup_plugHwi
was called again, the reference count was greater than 1,
and the Hwi would never be plugged again.

This issue manifests when Ipc_detach is called followed by
Ipc_attach.

This patch fixes the issue by checking against the appropriate
refcount for the mbxIdx.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>