tests: DRA7XX: Fix clock configuration in gatempapp.cfg
The gatempapp.cfg was not selecting the proper clock
config for DSP1. This patch fixes the gatempapp.cfg
file to use the same common Dsp_vayu.cfg and Dsp_vayu_power.cfg
files to do the common configurations including clock
configuration.

Without this, the watchdog will not be proper and will
result in unexpected watchdog timings for the messageq_fault
tests for DSP1.

Reported-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Tested-by: Suman Anna <s-anna@ti.com>