Resource: Rename the IpcMemory name in Resource.xs
The Resource module is derived from the IpcMemory module within
the sysbios-rpmsg codebase originally, but the Resource.xs file
is still using the IpcMemory name and so the config processing of
this module will print erroneous names on any misconfiguration.
Rename the variable IpcMemory to Resource like is normally done
with .xs files.

Signed-off-by: Suman Anna <s-anna@ti.com>