Remoteproc: DRA7xx: Adjust entries to fix carveout allocation failures
If the linux device tree uses DMA pools for carving out memory for the
slave cores, the DMA pools uses the next power of 2 page order
fitting the requested size (eg: 6 MB gets attempted to be allocated
using 8 MB and at 8 MB alignments), and this may result in allocation
failures if the overall DMA pool size does not have enough room to
support the different RSC_CARVEOUT aligned size.

This adjustment of the entries are updated carefully to make sure the
allocation does not fail due to this alignment. The 1 MB IPC_DATA
carveout entry is placed immediately after the vdev resource so that
the image carveouts are always aligned on an even 2 MB boundary.

Signed-off-by: Sam Nelson <sam.nelson@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>